requirements.txt
openai==1.14.1
python-dotenv==1.0.1
fastapi==0.110.0
starlette==0.36.3
uvicorn==0.28.0
click==8.1.7